Resume analyser/requirements.txt
flask
joblib
scikit-learn
numpy
scipy
pdfplumber
pytesseract
Pillow
gunicorn
